requirements.txt
flask==3.0.0
twilio==8.10.0
gunicorn==21.2.0
flask-sock==0.7.0
simple-websocket==1.0.0
openai>=1.3.0
nemo_toolkit[asr]==1.23.0
python-dotenv==1.0.0
librosa==0.10.1
soundfile==0.12.1
numpy==1.24.3
scipy==1.11.4
torch==2.1.2
elevenlabs>=1.0.0